+++ philo_actions.c
@@ -14,14 +14,14 @@
 
 void    philo_eat(t_philo *philo)
 {
-    double  current_timestamp;
+    long int  current_timestamp;
     
     //if (philo->num_forks == 2)
     if (philo->left_fork == taken && philo->right_fork == taken)
     {
         philo->state = eating;
         current_timestamp = get_timestamp();
-        printf("%f %d is eating\n", current_timestamp, philo->number);
+        printf("%ld %d is eating\n", current_timestamp, philo->number);
     }
     usleep((philo->time_to_eat)/1000);
     philo->time_last_meal = get_timestamp();
@@ -30,21 +30,21 @@ void    philo_eat(t_philo *philo)
 
 void    philo_sleep(t_philo *philo)
 {
-    double  current_timestamp;
+    long int  current_timestamp;
     
     philo->state = sleeping;
     current_timestamp = get_timestamp();
-    printf("%f %d is sleeping\n", current_timestamp, philo->number);
+    printf("%ld %d is sleeping\n", current_timestamp, philo->number);
     usleep((philo->time_to_sleep)/1000);
 }
 
 void    philo_think(t_philo *philo)
 {
-    double  current_timestamp;
+    long int  current_timestamp;
     
     philo->state = thinking;
     current_timestamp = get_timestamp();
-    printf("%f %d is thinking\n", current_timestamp, philo->number);
+    printf("%ld %d is thinking\n", current_timestamp, philo->number);
 }
 
 void    philo_take_fork(t_philo *philo, char *side)
@@ -56,7 +56,7 @@ void    philo_take_fork(t_philo *philo, char *side)
             philo->left_fork = taken;
             philo->next->right_fork = taken;
             philo->num_forks++;
-            printf("%f %d has taken its left fork\n", get_timestamp(), philo->number);
+            printf("%ld %d has taken its left fork\n", get_timestamp(), philo->number);
         }
     }
     else
@@ -66,16 +66,16 @@ void    philo_take_fork(t_philo *philo, char *side)
             philo->right_fork = taken;
             philo->prev->left_fork = taken;
             philo->num_forks++;
-            printf("%f %d has taken its right fork\n", get_timestamp(), philo->number);
+            printf("%ld %d has taken its right fork\n", get_timestamp(), philo->number);
         }
     }
 }
 
 void    philo_die(t_philo *philo)
 {
-    double  current_timestamp;
+    long int  current_timestamp;
     
     philo->state = dead;
     current_timestamp = get_timestamp();
-    printf("%f %d died\n", current_timestamp, philo->number);
+    printf("%ld %d died\n", current_timestamp, philo->number);
 }
